// SulambiVosaMain/Technology Transfer _ Sulambi VMS/Source Code/sulambi-frontend-main/sulambi-frontend-main/src/interface/types.ts
export interface SignatoriesType {
  preparedBy: string;
  reviewedBy: string;
  recommendingApproval1: string;
  recommendingApproval2: string;
  approvedBy: string;
  preparedTitle: string;
  reviewedTitle: string;
  approvedTitle: string;
  recommendingSignatory1: string;
  recommendingSignatory2: string;
}

export interface ExternalEventProposalType {
  id: number;
  extensionServiceType: string;
  eventProposalType: string;
  externalServiceType: string;
  title: string;
  location: string;
  durationStart: number;
  durationEnd: number;
  sdg: string;
  orgInvolved: string;
  programInvolved: string;
  projectLeader: string;
  partners: string;
  beneficiaries: string;
  totalCost: string;
  sourceOfFund: string;
  rationale: string;
  objectives: string;
  expectedOutput: string;
  description: string;
  financialPlan: string;
  dutiesOfPartner: string;
  evaluationMechanicsPlan: string;
  sustainabilityPlan: string;
  createdBy: any;
  status: "editing" | "submitted" | "accepted" | "rejected";
  toPublic: boolean;
  hasReport?: boolean;
  createdAt?: number;
  signatoriesId: SignatoriesType;
}

export interface InternalEventProposalType {
  id: number;
  title: string;
  venue: string;
  durationStart: number;
  durationEnd: number;
  modeOfDelivery: string;
  projectTeam: string;
  partner: string;
  participant: string;
  maleTotal: string;
  femaleTotal: string;
  rationale: string;
  objectives: string;
  description: string;
  workPlan: any;
  financialRequirement: any;
  evaluationMechanicsPlan: any;
  sustainabilityPlan: string;
  createdBy: any;
  status: "editing" | "submitted" | "accepted" | "rejected";
  toPublic: boolean;
  hasReport?: boolean;
  createdAt?: number;
  signatoriesId: SignatoriesType;
  eventProposalType?: string;
  activities?: Array<{ activity_name: string; months: number[] }>;
}